% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagadata.R
\docType{data}
\name{SAGA_Data}
\alias{SAGA_Data}
\title{SAGA Data}
\format{A 39428 x 92 data frame with post-op data. E}
\usage{
data(SAGA_Data)
}
\description{
These are the SAGA core data for model building. You'll need them loaded when using saga. Each of the 92 arrays provides 39428 probes with
array intensity information. Enjoy!
}
\examples{
data(SAGA_Data)
head(SAGA_Data)
}
\keyword{datasets}
